Create filter services to abstract some logic from filter component
